PO3D-VQA/6D_pose_estimator/requirements.txt
Pillow==7.0.0
matplotlib==3.1.3
numpy==1.18.1
BboxTools==1.0.2
torch==1.5.1
seaborn==0.10.0
scipy==1.4.1
opencv-python==4.2.0.34
git+https://github.com/facebookresearch/pytorch3d.git
